requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
databases==0.9.0
ecdsa==0.19.1
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
numpy==2.3.1
pandas==2.3.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==1.7.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
rsa==4.9.1
shim==0.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
